Runtime Error на LeetCode ошибка возникает на 31 строке, а именно minval.back() "ошибка 21: средство отчистки адресов: переполнение кучи-буфера по адресу 0x5020000001a8 на ПК 0x559404007a61 bp 0x7fff93685e10 sp 0x7fff93685e08".
у меня на компиляторе всё работает, на сайте нет.

+++ MyLibrary/Task5_MinStack/MinStack.h
@@ -3,8 +3,8 @@
 using namespace std;
 class MinStack {
 private:
-    vector<int> MyStack;
-    vector<int> minval;
+    vector<long int> MyStack;
+    vector<long int> minval;
     bool NewExit = true;
 public:
